pyproject.toml
[project]
name = "dayplanner-back"
version = "0.1"

dependencies = [
    "fastapi==0.124.2",
    "uvicorn",
    "psycopg2-binary",
    "sqlalchemy",
    "cryptography",
    "authlib",
    "google-api-python-client",
    "httpx",
    "python-jose[cryptography]"
]